Group and move factories into folder
