+++ config/common.php
@@ -7,8 +7,8 @@
  */
 
 return [
-    \Yiisoft\Rbac\RuleFactoryInterface::class => [
-        '__class' => \Yiisoft\Rbac\RuleFactory::class,
+    \Yiisoft\Rbac\Factories\RuleFactoryInterface::class => [
+        '__class' => \Yiisoft\Rbac\Factories\RuleFactory::class,
     ],
     \Yiisoft\Access\CheckAccessInterface::class => [
         '__class' => \Yiisoft\Rbac\DenyAll::class,
