+++ src/Factories/RuleFactory.php
@@ -6,19 +6,19 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace Yiisoft\Rbac;
+namespace Yiisoft\Rbac\Factories;
+
+use Yiisoft\Rbac\Rule;
 
 /**
  * @author Andrii Vasyliev <sol@hiqdev.com>
  *
  * @since 3.0
  */
-interface RuleFactoryInterface
+class RuleFactory implements RuleFactoryInterface
 {
-    /**
-     * @param string|mixed $name class name or other rule definition.
-     *
-     * @return Rule created rule.
-     */
-    public function create($name): Rule;
+    public function create(string $class): Rule
+    {
+        return new $class();
+    }
 }
