+++ src/Factories/RuleFactoryInterface.php
@@ -6,20 +6,21 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace Yiisoft\Rbac;
+namespace Yiisoft\Rbac\Factories;
+
+use Yiisoft\Rbac\Rule;
 
 /**
  * @author Andrii Vasyliev <sol@hiqdev.com>
  *
  * @since 3.0
  */
-class RuleFactory implements RuleFactoryInterface
+interface RuleFactoryInterface
 {
     /**
-     * {@inheritdoc}
+     * @param string $name class name or other rule definition.
+     *
+     * @return Rule created rule.
      */
-    public function create($class): Rule
-    {
-        return new $class();
-    }
+    public function create(string $name): Rule;
 }
